+++ CHANGELOG.md
@@ -1,3 +1,39 @@
+# BetterBlizzFrames 1.9.3
+## Midnight
+### New
+- Add Jazggz profile (www.twitch.tv/jazggz). Thank you for sharing <3
+- New "Improved Target Highlight" setting for PartyFrames (General).
+- New "Hide Dispel Icon" sub-setting for "Hide Dispel Overlay" for PartyFrames (General).
+- New "Change ActionBar CD Size" setting (Misc).
+- New "Hide All Manabar Text" setting (Misc).
+### Tweak
+- Interrupt Icon no longer shows up on un-interruptible casts like how it used to before Midnight.
+- Add few more Edit Mode stuff to the Edit Mode transparancy slider.
+- Change border style a bit on PartyFrames due to a secret issue in Blizzards code.
+### Bugfix
+- Fix secret error issue due to aura borders on partyframe doing some lua errors.
+- Temporary fix with debug print on secret error on party arena names settings. Pls report if you see it.
+
+# BetterBlizzFrames 1.9.2c
+## Midnight
+### Tweak
+- Update Gladiator/Strategist/Legend achievements for Glad Tracker.
+- Potential fix for a secret error.
+- Update Mes Profile (www.twitch.tv/notmes). Thank you for sharing.
+
+# BetterBlizzFrames 1.9.2b
+## Midnight
+### New
+- Buffs & Debuffs: New "Only mine" setting for Show cooldown text on auras. Will only show cd text on your own auras to reduce some clutter. This is now enabled by default.
+### Tweak
+- DarkMode: Auras now also skins PartyFrame Buffs (Raid-Style). Does not do raid atm.
+- Fix FocusFrame Castbar position (with aura settings enabled and normal castbar behaviour) being 5 pixels too low compared to TargetFrame castbar. They're now identical as they should be.
+- Move the Big Debuffs setting from Misc to Extra Features topright in /bbf.
+- Fix position of "Unconscious" text on healthbar for Classic Frames and No Portait settings.
+- Fix a potential error Mini-PlayerFrame ran in combat during login.
+### Localization
+- Add Simplified Chinese localization by LDZhou @ GitHub. Thank you!
+
 # BetterBlizzFrames 1.9.2
 ## Midnight
 ### New